share sft dummy renderer fixture
